+++ drivers/include/nrfx_rng.h
@@ -49,12 +49,10 @@ typedef void (* nrfx_rng_evt_handler_t)(uint8_t rng_data);
  * @brief Function for initializing the nrfx_rng module.
  *
  * @param[in]  p_config Pointer to the structure with initial configuration.
- * @param[in]  handler  Event handler provided by the user. Handler is required
- *                      to work with the module. NULL value is not supported.
+ * @param[in]  handler  Event handler provided by the user. Must not be NULL.
  *
  * @retval  NRFX_SUCCESS                   Driver was successfully initialized.
  * @retval  NRFX_ERROR_ALREADY_INITIALIZED Driver was already initialized.
- * @retval  NRFX_ERROR_INVALID_PARAM       Handler value is NULL.
  */
 nrfx_err_t nrfx_rng_init(nrfx_rng_config_t const * p_config, nrfx_rng_evt_handler_t handler);