+++ drivers/include/nrfx_spis.h
@@ -139,8 +139,10 @@ typedef void (*nrfx_spis_event_handler_t)(nrfx_spis_event_t event);
  * @param[in] p_instance    Pointer to the driver instance structure.
  * @param[in] p_config      Pointer to the structure with initial configuration.
  * @param[in] event_handler Function to be called by the SPI slave driver upon event.
+ *                          Must not be NULL.
  *
  * @retval NRFX_SUCCESS             If the initialization was successful.
+ * @retval NRFX_ERROR_INVALID_STATE If the instance is already initialized.
  * @retval NRFX_ERROR_INVALID_PARAM If an invalid parameter is supplied.
  * @retval NRFX_ERROR_BUSY          If some other peripheral with the same
  *                                  instance ID is already in use. This is