+++ drivers/src/nrfx_rng.c
@@ -22,14 +22,10 @@ static nrfx_rng_evt_handler_t m_rng_hndl;
 nrfx_err_t nrfx_rng_init(nrfx_rng_config_t const * p_config, nrfx_rng_evt_handler_t handler)
 {
     NRFX_ASSERT(p_config);
+    NRFX_ASSERT(handler);
     if (m_rng_state != NRFX_DRV_STATE_UNINITIALIZED)
     {
         return NRFX_ERROR_ALREADY_INITIALIZED;
-    }
-
-    if (handler == NULL)
-    {
-        return NRFX_ERROR_INVALID_PARAM;
     }
 
     m_rng_hndl = handler;